Avoid re-building on local machine

+++ build.sh
@@ -1,5 +1,12 @@
 #! bash
 
+[[ -e MagiskFonts.zip ]] && rm MagiskFonts.zip
+if [[ -z $TERMUX_VERSION ]]; then
+	echo "Skip building font-scan if not in termux"
+	7z a MagiskFonts.zip META-INF customize.sh module.prop system tool
+	exit
+fi
+pushd fontconfig
 pushd fontconfig
 rm -rf build
 meson build
